requirements.txt
Django==3.1.5
djangorestframework==3.12.2
requests==2.25.1
environs~=9.3.0